Add reuse support to Owner protocol and samples
Adding support for basic Reuse protocol in the Owner storage sample and
protocol.
- New methods have been added into the To2ServerStorage interface
to facilitate this. OwnerDbStorage has been updated to provide
implementations.
- To2ServerService has been updated to handle Reuse scenario.

TO-DOs: Throw an exception if TO2Client sends the replacement HMAC not
equal to - a valid HMAC, or, CBOR NULL, only after TO2Client is updated
for reuse/resale scenarion.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>