Fix: fetch the correct EC curve name when verifying
When EC parameter field toString() is called on bouncycastle
ECNamedCurveSpec objects, it produces a java-style "class@address"
string, which is not what we want.

Instead, fetch and test the curve name field from the parameter
spec object if it's a bouncycastle object.

Accept the legacy name for secp256r1: prime256v1.

Signed-off-by: Greg Bean <greg.bean@intel.com>