Adding Serviceinfo marshaller and owner SI flow
Adding Serviceinfo marshaller that packs serviceinfo to be sent across
multiple messages. It fetches the list of serviceinfo entries and
combines multiple entries together, if they can fit as per the max MTU
limit. If any entry cannot be fit in the message as per the MTU, it is
partitioned into separate smaller messages, if they support partitioning
(splitting).
The serviceinfo fetch/store operations are added in the Owner TO2
storage sample. The database toble to store serviceinfo contains a
column that specifies the type of data:
- CBOR: for CBOR encoded primitive data that will be transferred to
the device as is (no splitting), and
- Plain: for regular binary data that will allow splitting of content
to be sent across messages

Though the serviceinfo implementation is generic, to maintain the
transfer flow of SDO_SYS module (filedesc/write/exec), the serviceinfo
for any device is fetched/sent in their insertion order as defined by a
timestamp column.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>