Owner and Device ServiceInfo Implementation. (#30)
Adding ServiceInfo marshaller that packs ServiceInfo to be sent across
multiple messages. It fetches the list of ServiceInfo entries and
combines multiple entries together, if they can fit as per the max MTU
limit. If any entry cannot be fit in the message as per the MTU, it is
partitioned into separate smaller messages, if they support partitioning
(splitting).

The ServiceInfo fetch/store operations are added in the Owner TO2
storage sample. The database table to store ServiceInfo contains a
column that specifies the type of data:
- CBOR: for CBOR encoded primitive data that will be transferred to
the device as is (no splitting), and
- Plain: for regular binary data that will allow splitting of content
to be sent across messages

Though the ServiceInfo implementation is generic, to maintain the
transfer flow of SDO_SYS module (filedesc/write/exec), the ServiceInfo
for any device is fetched/sent in their insertion order as defined by a
timestamp column.

With respect to Device ServiceInfo implementation,

- The device ServiceInfo fetch/store operations are added in the Device
Component Samples. The ServiceInfo data (CBOR encoded type and regular
binary data) will be sent to Owner without splitting.
- Added DeviceServiceInfoSequence which serializes data in the format
required for ServiceInfoEntry.
- Updated To2StorageTest class with device service info.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>
Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>