serviceinfo: Optimize transfer rate (#107)
- The earlier approach created the ServiceInfo messages from the start
index until the service info that the Owner wishes to send was found,
while iterating over it. This meant that more the number of ServiceInfo
messages to be sent, the longer it took to iterate and find the message
owner needed to send.
- In the current approach that fixes the issue, the
'ServiceinfoMarshaller' is serialized/de-serialized to-and-from
database. This allows the Owner to directly find the message and send,
since the Marshaller state is stored, and the iteration to create
messages is completely avoided.
- Additionally, fixing an issue with the Device where in it sent the same
DeviceServiceinfo message in Type 68 again and again to the owner.

Signed-off-by: Chandrakar, Prateek <prateek.chandrakar@intel.com>