SVI Grouping
Eliminating serviceinfo one on one mapping to each device.
Devicetype identification is done based on the criteria specified for
each device type. This critera(s) shoudl map with the DSI keys sent by
the device. If the expected value specified matches with the actual
value for the DSI keys in all criteria of the device, the corresponding
owner SVI string will be used for owner SVI transfer.

The first match is used as the devicetype, so device with restrictive criteria
should be specified before device with less restrictive criteria. If no
criteria is specified the device type will default to 'default' type. If
user wants to opt out of SVI transer specify SVI string as null.

Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>