fix voucher validation
Voucher leaf certificate should be the first certificate
(0) instead of certificate 1. After fixing this issue there
is no longer any need for a specific check for OnDie so the
corresponding code can be removed.

In addition, code to validate the root CA is OnDie is moved
to the OnDie signature validation function.

Signed-off-by: Easterday, John <john.easterday@intel.com>