SVI Grouping (#148)
Eliminating one-on-one mapping of ServiceInfo for each device.

Devicetype identification is done based on the criteria specified for
each device type. This criteria(s) should map with the DSI keys sent by
the device. If the expected value specified matches with the actual
value for the DSI keys in all criteria of the device, the corresponding
owner SVI string will be used for owner SVI transfer.

The first match is used as the devicetype, so device with restrictive criteria
should be specified before device with less restrictive criteria. If no
criteria is specified, 'default' devicetype is used. If user wants to opt out
of SVI transfer, SVI string should be specified as null.

Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>