Resale scenario using reseller key for creating size 0 voucher.
Resale is enabled when either or all of Replacement GUID, Replacement RV Info
or Replacement owner key are not the same as the existing value.

Until now, the implementation was enabling resale only with the change
in GUID and RV Info but with this feature, user can provide owner2 key that
will be used for creating the new size 0 voucher, instead of using
existing owner key.

This implementation creates a special case scenario when we use reseller key
as owner2 key and eliminates the need to add the previous owner key in reseller
keystore as the device ownership will already be with the reseller upon generation
of size 0 voucher.

Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>