Fixing EC384 support
PRI Manufacturer was picking up EC256 certificate for the issuer certchain
while creating device ownership voucher for both EC256 and EC384 device
key types. This was resulting in onboarding failure during TO2 due to hash mismatch.
This was happening because in the Manufacturer Storage method `getCertChain`,
only the key algorithm was being checked, which is `ECDH` for both EC256 and EC384
keytypes.

To resolve this issue, an additional check has been added to the
function and the keytype check has been added on top of existing checks
to retrieve the expected certificate.

Signed-off-by: Darshini Parikh <darshini.k.parikh@intel.com>