Isolate test credentials from demo folder
Providing the test credentials to run a demo raises security concern as
there is a potential of these test credentials being used in production
deployments.

This is the first step to remove the test credentials from demo folder.
In a subsequent PR, the test credentials generation would be automated
(using component-samples/scripts/keys_gen.sh script).

Signed-off-by: Behera, Tushar Ranjan <tushar.ranjan.behera@intel.com>