Remove PKCS11 support and make docker scripts uniform
For manufacturer and reseller components, the PKCS12 keystores are
converted to PKCS11 format and imported into SoftHSM to simulate HSM
handling. The current implementation requires the manufacturer and
reseller docker instances to run with ROOT user.

Considering we are providing example implementations in
component-samples, it is better to keep all the docker configurations
similar. This would help end-user to launch all the containers in a
uniform manner.

After removal of the PKCS11 support, a non-ROOT user `fdo` is created in
all the docker containers as per guidance from security best practices.

With these changes, the permission of the SSL keystores/truststores can
be set to read-only.

Signed-off-by: Behera, Tushar Ranjan <tushar.ranjan.behera@intel.com>